requirements.txt
customtkinter>=5.2.2
pillow>=10.0
python-nmap>=0.7.1
pystray>=0.19.4
